pyproject.toml
[project]
name = "pythontamplate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fusepy>=3.0.1",
    "msgspec>=0.19.0",
]

[tool.ruff]
extend-exclude = [".venv"]
line-length = 100
preview = true

[tool.ruff.lint]
explicit-preview-rules = true
ignore = [
    "S101",
    "RUF001",
    "RUF002",
    "RUF003",
    "RUF012",
    "PD901",
    "B905",
    "TRY003",
    "PLR0912",
    "PLR0913",
    "PLR0915",
    "PLR2004",
    "RET504",
    "S603",
    "S606",
    "A002",
    "A001",
    "BLE001",
    "PLC0414",
]
select = [
    "E",
    "F",
    "W",
    "I",
    "UP",
    "ASYNC",
    "S",
    "BLE",
    "FBT003",
    "B",
    "A",
    "COM818",
    "C4",
    "ISC",
    "PIE",
    "PYI",
    "PT",
    "Q",
    "RSE",
    "RET",
    "SLOT",
    "SIM",
    "TC",
    "PTH",
    "PD",
    "PL",
    "NPY",
    "FAST",
    "AIR",
    "RUF",

]
extend-unsafe-fixes = []
preview = true
[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true
ignore-fully-untyped = true
mypy-init-return = true
suppress-none-returning = true

[tool.ruff.lint.flake8-type-checking]
exempt-modules = ["typing", "typing_extensions"]
quote-annotations = true

[tool.ruff.lint.pydocstyle]
convention = "pep257"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint.flake8-implicit-str-concat]
allow-multiline = false

[tool.pyright]
exclude = ["venv", ".venv", "python"]
typeCheckingMode = "standard"

analyzeUnannotatedFunctions = true
deprecateTypingAliases = true
strictDictionaryInference = true
strictListInference = true
strictParameterNoneValue = true
strictSetInference = true
reportIncompatibleVariableOverride = false
reportAbstractUsage = true
reportAny = false
reportArgumentType = true
reportAssertAlwaysTrue = true
reportAssertTypeFailure = true
reportAssignmentType = true
reportAttributeAccessIssue = true
reportCallInDefaultInitializer = true
reportCallIssue = true
reportConstantRedefinition = true
reportDeprecated = true
reportDuplicateImport = "warning"
reportFunctionMemberAccess = true
reportGeneralTypeIssues = true
reportImplicitOverride = "information"
reportImplicitRelativeImport = false
reportImplicitStringConcatenation = "warning"
reportImportCycles = false
reportIncompleteStub = true
reportInconsistentConstructor = "warning"
reportIndexIssue = true
reportInvalidCast = true
reportInvalidStringEscapeSequence = true
reportInvalidStubStatement = true
reportInvalidTypeArguments = true
reportInvalidTypeForm = true
reportInvalidTypeVarUse = "information"
reportMatchNotExhaustive = true
reportMissingImports = "information"
reportMissingModuleSource = true
reportMissingParameterType = false
reportMissingSuperCall = false
reportMissingTypeArgument = false
reportMissingTypeStubs = false
reportNoOverloadImplementation = true
reportOperatorIssue = true
reportOptionalCall = true
reportOptionalContextManager = true
reportOptionalIterable = true
reportOptionalMemberAccess = true
reportOptionalOperand = true
reportOptionalSubscript = true
reportOverlappingOverload = true
reportPrivateImportUsage = "warning"
reportPrivateLocalImportUsage = false
reportPrivateUsage = "warning"
reportPropertyTypeMismatch = true
reportRedeclaration = true
reportReturnType = true
reportSelfClsParameterName = true

reportTypeCommentUsage = true
reportTypedDictNotRequiredAccess = true
reportUnboundVariable = "warning"
reportUndefinedVariable = true
reportUnhashable = true
reportUninitializedInstanceVariable = false
reportUnknownArgumentType = false
reportUnknownLambdaType = false
reportUnknownMemberType = false
reportUnknownParameterType = false
reportUnknownVariableType = false
reportUnnecessaryCast = "warning"
reportUnnecessaryComparison = "warning"
reportUnnecessaryContains = "warning"
reportUnnecessaryIsInstance = "warning"
reportUnnecessaryTypeIgnoreComment = "warning"
reportUnreachable = true
reportUnsafeMultipleInheritance = false
reportUnsupportedDunderAll = true
reportUntypedBaseClass = true
reportUntypedClassDecorator = true
reportUntypedFunctionDecorator = true
reportUntypedNamedTuple = true
reportUnusedCallResult = false
reportUnusedClass = "warning"
reportUnusedCoroutine = "warning"
reportUnusedExcept = "warning"
reportUnusedExpression = "warning"
reportUnusedFunction = "warning"
reportUnusedImport = false
reportUnusedVariable = false

reportWildcardImportFromLibrary = true
[tool.coverage.report]
ignore_errors = true
skip_covered = true
skip_empty = true
[tool.coverage.run]
branch = true
concurrency = ["multiprocessing"]
parallel = true
[tool.coverage.lcov]
line_checksums = true
